oe-core: Using fixed hash.
Yesterday openembedded-core upgraded GCC to 6.4 and backported a bunch
of patches. This breaks compilation of the 6-args hypercall
implementation for V4V with the following error:
> error: bp cannot be used in asm here

This is because the 6-args HC uses ebp to pass the last argument for
x86_32 and default with -O2 seems to be -fno-omit-frame-pointer.

A quick and dirty work-around would be to use -fomit-frame-pointer...
All in all this is yet another reason, if one is still needed, that V4V
needs that refactoring we all want.

Use a fixed revision of openembedded-core just before the upgrade until
this is sorted out.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>